Fix Multiple user Stack component
